refactor: move base script to client script
